Add VERBOSE; start requests at #0; add info to body.
A boolean VERBOSE at the top of the source file now
allows to switch between VERBOSE and non-VERBOSE mode.

Request are now numbered starting from #0, so their number
matches the index into the handles[] and headers[] arrays.
Note that the server still starts counting replies at 1,
so there is no longer a match between replies and request
numbers, but that wasn't the case anyway after a new
connection is started.

The html body sent back by the server now contains the
connection, reply and request numbers.